requirements.txt
flask
pycryptodome
datetime
requests
uuid
argparse
cryptography